:sparkles:feat(e10_t10) implements cover img upload
