+++ docker-compose.dev.yml
@@ -69,8 +69,6 @@ services:
 
   frontend:
     container_name: frontend
-    env_file:
-      - .env  
     build:
       context: .
       dockerfile: Dockerfile.dev
