+++ src/utils/interfaces/Interfaces.ts
@@ -6,6 +6,7 @@ export interface IBlog {
   template: number
   createdAt: Date
   user: IUser
+  coverUrl: string | null
   articles: IArticle[]
 }
